backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
celery[redis]
redis
aiohttp
playwright
beautifulsoup4
lxml
transformers
torch
spacy
deepface
opencv-python
Pillow
weasyprint
jinja2
python-jose[cryptography]
passlib[bcrypt]
cryptography
python-telegram-bot
python-dotenv
requests
pytest
flake8
mypy
bandit